FABN-1570: Explicit dependency versions on publishing (#246)
Dependencies on other packages within the Node SDK now use the exact published package versions instead of the npm tag.

The publishing script uses the version and tag from the top-level package.json only. Versions and tags of individual packages are rewritten to reflect the top-level package.json values at publish time. Unstable releases must have a pre-release version number (including a "-") and the pre-release version is incremented from the previously published pre-release package version. Stable releases must specify a release version, and the specified version is used without modification.

Packages with a tag of "skip" in their local package.json will not be published. Dependencies on those packages will be rewritten to the previously published version. This might be useful to recover from a situation where an integration build failed after publishing some but not all of the packages for a stable release. Generally all modules should be published together and the skip capability should not be used.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>